communicator/docker: Refactor for docker API
This commit switches the the Docker communicator to use the Docker
API, via go-dockercilent.

The functionality makes full use of the API. Upload, Download,
UploadDir, and DownloadDir have been set up to work with the tar file
upload/download scheme the API uses, and the use of the Exec API has
allowed Start() to be heavily simplified.

This should address a long-standing issue with the Docker
communicator where it hung indefinitely during docker-inside-docker
(nested) or docker-outside-of-docker (sibling) container provisioning.
It also adds DownloadDir to the docker communicator.

References:

 mitchellh/pakcer#3102
 mitchellh/packer#3916